requirements.txt
Flask
gunicorn
google-cloud-vision